+++ x/emissions/module/rewards/scores_test.go
@@ -1091,7 +1091,6 @@ func GenerateWorkerDataBundles(s *RewardsTestSuite, blockHeight int64, topicId u
 
 func GenerateMoreInferencesDataBundles(s *RewardsTestSuite, blockHeight int64, topicId uint64) []*types.WorkerDataBundle {
 	var newInferences []*types.WorkerDataBundle
-	oldForecaster := s.addrs[5]
 	worker1Addr := s.addrs[10]
 	worker2Addr := s.addrs[11]
 
@@ -1105,7 +1104,7 @@ func GenerateMoreInferencesDataBundles(s *RewardsTestSuite, blockHeight int64, t
 		Forecast: &types.Forecast{
 			TopicId:     topicId,
 			BlockHeight: blockHeight,
-			Forecaster:  oldForecaster.String(),
+			Forecaster:  worker1Addr.String(),
 			ForecastElements: []*types.ForecastElement{
 				{
 					Inferer: s.addrs[7].String(),
@@ -1140,7 +1139,7 @@ func GenerateMoreInferencesDataBundles(s *RewardsTestSuite, blockHeight int64, t
 		Forecast: &types.Forecast{
 			TopicId:     topicId,
 			BlockHeight: blockHeight,
-			Forecaster:  oldForecaster.String(),
+			Forecaster:  worker2Addr.String(),
 			ForecastElements: []*types.ForecastElement{
 				{
 					Inferer: s.addrs[5].String(),
@@ -1170,8 +1169,6 @@ func GenerateMoreInferencesDataBundles(s *RewardsTestSuite, blockHeight int64, t
 
 func GenerateMoreForecastersDataBundles(s *RewardsTestSuite, blockHeight int64, topicId uint64) []*types.WorkerDataBundle {
 	var newForecasts []*types.WorkerDataBundle
-	oldInferencer1 := s.addrs[5]
-	oldInferencer2 := s.addrs[6]
 	worker1Addr := s.addrs[10]
 	worker2Addr := s.addrs[11]
 
@@ -1179,7 +1176,7 @@ func GenerateMoreForecastersDataBundles(s *RewardsTestSuite, blockHeight int64, 
 		Inference: &types.Inference{
 			TopicId:     topicId,
 			BlockHeight: blockHeight,
-			Inferer:     oldInferencer1.String(),
+			Inferer:     worker1Addr.String(),
 			Value:       alloraMath.MustNewDecFromString("0.01251"),
 		},
 		Forecast: &types.Forecast{
@@ -1198,15 +1195,15 @@ func GenerateMoreForecastersDataBundles(s *RewardsTestSuite, blockHeight int64, 
 			},
 		},
 	}
-	worker1Sig, err := GenerateWorkerSignature(s, worker1InferenceForecastBundle, oldInferencer1)
+	worker1Sig, err := GenerateWorkerSignature(s, worker1InferenceForecastBundle, worker1Addr)
 	s.Require().NoError(err)
 	worker1Bundle := &types.WorkerDataBundle{
-		Worker:                             oldInferencer1.String(),
+		Worker:                             worker1Addr.String(),
 		Nonce:                              &types.Nonce{BlockHeight: blockHeight},
 		TopicId:                            topicId,
 		InferenceForecastsBundle:           worker1InferenceForecastBundle,
 		InferencesForecastsBundleSignature: worker1Sig,
-		Pubkey:                             GetAccPubKey(s, oldInferencer1),
+		Pubkey:                             GetAccPubKey(s, worker1Addr),
 	}
 	newForecasts = append(newForecasts, worker1Bundle)
 
@@ -1214,7 +1211,7 @@ func GenerateMoreForecastersDataBundles(s *RewardsTestSuite, blockHeight int64, 
 		Inference: &types.Inference{
 			TopicId:     topicId,
 			BlockHeight: blockHeight,
-			Inferer:     oldInferencer2.String(),
+			Inferer:     worker2Addr.String(),
 			Value:       alloraMath.MustNewDecFromString("0.01251"),
 		},
 		Forecast: &types.Forecast{
@@ -1233,15 +1230,15 @@ func GenerateMoreForecastersDataBundles(s *RewardsTestSuite, blockHeight int64, 
 			},
 		},
 	}
-	worker2Sig, err := GenerateWorkerSignature(s, worker2InferenceForecastBundle, oldInferencer2)
+	worker2Sig, err := GenerateWorkerSignature(s, worker2InferenceForecastBundle, worker2Addr)
 	s.Require().NoError(err)
 	worker2Bundle := &types.WorkerDataBundle{
-		Worker:                             oldInferencer2.String(),
+		Worker:                             worker2Addr.String(),
 		Nonce:                              &types.Nonce{BlockHeight: blockHeight},
 		TopicId:                            topicId,
 		InferenceForecastsBundle:           worker2InferenceForecastBundle,
 		InferencesForecastsBundleSignature: worker2Sig,
-		Pubkey:                             GetAccPubKey(s, oldInferencer2),
+		Pubkey:                             GetAccPubKey(s, worker2Addr),
 	}
 	newForecasts = append(newForecasts, worker2Bundle)